Move networking module into kernel
