+++ crates/kernel/src/networking/utils/arp_cache.rs
@@ -1,7 +1,9 @@
 use alloc::collections::BTreeMap;
-use kernel::device::system_time;
+use crate::device::system_timer;
 
-use crate::repr::{EthernetAddress, Ipv4Address};
+use crate::networking::repr::{EthernetAddress, Ipv4Address};
+// use crate::networking::utils::arp_cache::Timestamp;
+use core::time::Duration;
 
 // Replace std::time::Instant with a custom timestamp (e.g., u64 milliseconds)
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
@@ -36,32 +38,33 @@ impl ArpCache {
     pub fn set_eth_addr_for_ip(&mut self, ipv4_addr: Ipv4Address, eth_addr: EthernetAddress) {
         self.expire_eth_addr();
 
+        let now = system_timer::get_time(); // Use system_time::get_time() to get current time
         if self.entries.is_empty() {
-            self.in_cache_since_min = now;
+            self.in_cache_since_min = Timestamp(now);
         }
 
         self.entries.insert(
             ipv4_addr,
             Entry {
                 eth_addr,
-                in_cache_since: now,
+                in_cache_since: Timestamp(now),
             },
         );
     }
 
     fn expire_eth_addr(&mut self) {
-        let now = system_time::get_time(); // Use system_time::get_time() to get current time
+        let now = system_timer::get_time(); // Use system_time::get_time() to get current time
 
         // If the cache has been in use for longer than the expiration period
-        if now > self.in_cache_since_min + self.expiration {
+        if now > self.in_cache_since_min.0 + (self.expiration.as_micros() as u64) {
             let expiration = self.expiration;
-            self.entries.retain(|_, entry| now.saturating_sub(entry.in_cache_since) <= expiration);
+            self.entries.retain(|_, entry| now.saturating_sub(entry.in_cache_since.0) <= (expiration.as_micros() as u64));
 
             // Update the minimum in_cache_since timestamp
             let in_cache_since = self.entries.iter().map(|(_, entry)| entry.in_cache_since);
             self.in_cache_since_min = match in_cache_since.clone().min() {
                 Some(min_time) => min_time,
-                None => now,
+                None => Timestamp(now),
             }
         }
     }
