+++ crates/kernel/src/networking/utils/checksum.rs
