+++ crates/kernel/src/networking/utils/mod.rs
