+++ crates/kernel/src/networking/utils/range.rs
